remove legacy code and clean up #12
