+++ app/(protected)/layout.tsx
@@ -1,0 +1,28 @@
+import { authOptions } from "@/auth";
+import { AppSidebar } from "@/components/app-sidebar";
+import { SidebarInset, SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
+import { Separator } from "@radix-ui/react-separator";
+import { getServerSession } from "next-auth";
+import { redirect } from 'next/navigation'
+import { ReactNode } from "react";
+
+
+export default async function Layout({ children }: { children: ReactNode }) {
+  const session = await getServerSession(authOptions)
+  if (!session) return redirect("/login")
+
+  return (<SidebarProvider>
+    <AppSidebar />
+    <SidebarInset>
+      <header className="flex h-16 shrink-0 items-center gap-2">
+        <div className="flex items-center gap-2 px-4">
+          <SidebarTrigger className="-ml-1" />
+          <Separator orientation="vertical" className="mr-2 h-4" />
+        </div>
+      </header>
+      <div className="flex flex-1 flex-col gap-4 p-4 pt-0">
+        {children}
+      </div>
+    </SidebarInset>
+  </SidebarProvider>);
+}
