+++ app/(protected)/page.tsx
@@ -1,0 +1,8 @@
+
+
+export default async function page() {
+
+  return (<div className="w-full sm:p-6">
+    <h1>dashboard</h1>
+  </div>)
+}
