--- app/Producthunt.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const Producthunt = () => {
-  return (<a href="https://www.producthunt.com/posts/dcup-2?embed=true&utm_source=badge-featured&utm_medium=badge&utm_souce=badge-dcup&#0045;2" target="_blank"><img src="https://api.producthunt.com/widgets/embed-image/v1/featured.svg?post_id=880190&theme=light&t=1739712648146" alt="Dcup - Transform&#0032;messy&#0032;documents&#0032;into&#0032;clean&#0044;&#0032;structured&#0032;JSON | Product Hunt" style={{width: "250px", height: "54px" }} width="250" height="54" /></a>
-  )
-}
