--- app/robots.ts
@@ -1,12 +1,0 @@
-import type { MetadataRoute } from 'next'
- 
-export default function robots(): MetadataRoute.Robots {
-  return {
-    rules: {
-      userAgent: '*',
-      allow: '/',
-      disallow: ['/dashboard/', '/login/', '/integration/'],
-    },
-    sitemap: 'https://dcup.dev/sitemap.xml',
-  }
-}
