--- components/HeroImage/heroImage.module.css
@@ -1,268 +1,0 @@
-/* Global CSS Variables & Reset */
-.heroContainer {
-  --primary: #6366f1;
-  --secondary: #4f46e5;
-  --accent: #8b5cf6;
-  --background: #1a1a1a;
-  --json-bg: #0f172a;
-  --json-color: #7dd3fc;
-  --messy-bg: #fff;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-family: 'Inter', sans-serif;
-  overflow: hidden;
-  padding: 20px;
-}
-
-/* Main container - Desktop layout */
-.container {
-  position: relative;
-  width: 800px;
-  height: 500px;
-  border-radius: 12px;
-  overflow: hidden;
-}
-
-/* Messy data panel */
-.messy {
-  text-align: left;
-  position: absolute;
-  top: 50px;
-  left: 50px;
-  width: 300px;
-  height: 400px;
-  background: var(--messy-bg);
-  box-shadow: 0 10px 40px rgba(0, 0, 0, 0.5);
-  border-radius: 8px;
-  padding: 20px;
-  font-family: 'Courier New', monospace;
-  font-size: 13px;
-  color: #444;
-  overflow: hidden;
-  animation: messyOut 6s infinite ease-in-out;
-}
-
-.messy::before {
-  content: "%%%@#*^&!! ?-- /\\ messy data...\A<?xml version=\"1.0\"?>\A<!DOCTYPE note SYSTEM \"Note.dtd\">\A<note>\A  <to>User</to>\A  <from>Server</from>\A  <heading>Disorganized Data</heading>\A  <body>#!@$$%*&^ Unexpected content... 1234567890</body>\A</note>\A/* Lorem ipsum dolor sit amet, consectetur adipiscing elit messy Data need to clean with Dcup... */";
-  white-space: pre-wrap;
-}
-
-/* JSON data panel */
-.json {
-  position: absolute;
-  top: 50px;
-  right: 50px;
-  width: 400px;
-  height: 400px;
-  background: var(--json-bg);
-  border-radius: 8px;
-  padding: 25px;
-  color: var(--json-color);
-  font-family: 'Fira Code', monospace;
-  font-size: 14px;
-  overflow: auto;
-  opacity: 0;
-  transform: translateY(30px);
-  animation: jsonIn 6s infinite ease-in-out;
-}
-
-.json pre {
-  white-space: pre-wrap;
-  text-align: left;
-}
-
-/* Magic Wand Animation */
-.magicWand {
-  position: absolute;
-  top: 40%;
-  left: -100px;
-  font-size: 40px;
-  animation: wandMove 6s infinite ease-in-out;
-  z-index: 2;
-  filter: drop-shadow(0 0 8px rgba(255, 215, 0, 0.5));
-}
-
-.magicWand::before {
-  content: '🪄';
-  position: absolute;
-  animation: wandTwinkle 0.6s infinite alternate;
-}
-
-/* Keyframes */
-@keyframes wandMove {
-
-  0%,
-  40% {
-    transform: translateX(0) rotate(0deg);
-    opacity: 0;
-  }
-
-  45% {
-    opacity: 1;
-  }
-
-  50%,
-  60% {
-    transform: translateX(800px) rotate(720deg);
-  }
-
-  65% {
-    opacity: 1;
-  }
-
-  70%,
-  100% {
-    transform: translateX(1000px) rotate(900deg);
-    opacity: 0;
-  }
-}
-
-@keyframes wandTwinkle {
-  from {
-    filter: drop-shadow(0 0 8px rgba(255, 215, 0, 0.5));
-  }
-
-  to {
-    filter: drop-shadow(0 0 15px rgba(255, 215, 0, 0.8));
-  }
-}
-
-@keyframes messyOut {
-
-  0%,
-  40% {
-    opacity: 1;
-    transform: translateX(0) rotate(0deg);
-  }
-
-  45% {
-    opacity: 1;
-    transform: translateX(-30px) rotate(-2deg);
-  }
-
-  50% {
-    opacity: 0;
-    transform: translateX(-50px) rotate(-5deg);
-  }
-
-  95% {
-    opacity: 0;
-  }
-
-  100% {
-    opacity: 1;
-    transform: translateX(0) rotate(0deg);
-  }
-}
-
-@keyframes jsonIn {
-
-  0%,
-  45% {
-    opacity: 0;
-    transform: translateY(30px) scale(0.9);
-    filter: blur(5px);
-  }
-
-  50% {
-    opacity: 1;
-    transform: translateY(0) scale(1);
-    filter: blur(0);
-  }
-
-  90% {
-    opacity: 1;
-  }
-
-  95% {
-    opacity: 0;
-    transform: translateY(30px) scale(0.9);
-  }
-
-  100% {
-    opacity: 0;
-  }
-}
-
-/* Enhanced JSON Styling */
-.json {
-  border: 1px solid rgba(99, 102, 241, 0.3);
-  background: linear-gradient(145deg, #0f172a 0%, #1e293b 100%);
-  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.3);
-}
-
-.json::before {
-  content: 'formatted.json';
-  position: absolute;
-  top: 10px;
-  left: 15px;
-  color: #64748b;
-  font-size: 12px;
-  font-family: 'Inter', sans-serif;
-}
-
-.json-key {
-  color: #f472b6;
-}
-
-.json-string {
-  color: #7dd3fc;
-}
-
-.json-number {
-  color: #34d399;
-}
-
-/* Responsive Styles: Mobile Layout */
-@media (max-width: 768px) {
-
-  /* Convert container to a vertical, scrollable layout */
-  .container {
-    width: 100%;
-    height: auto;
-    padding: 20px;
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-    /* Remove overflow hidden to let content scroll if needed */
-    overflow: visible;
-    position: relative;
-  }
-
-  /* Reset the absolute positioning for panels */
-  .messy,
-  .json,
-  .magicWand {
-    position: relative;
-    left: 0;
-    top: 0;
-    transform: none;
-  }
-
-  /* Stack panels vertically */
-  .messy,
-  .json {
-    width: 100%;
-    height: auto;
-    animation: none;
-    /* Optionally disable or simplify animations on mobile */
-  }
-
-  /* Optionally keep a simple fade-in animation */
-  .messy {
-    animation: messyOut 6s infinite ease-in-out;
-  }
-
-  .json {
-    animation: jsonIn 6s infinite ease-in-out;
-  }
-
-  /* Place the magic wand at the top (or hide if not desired) */
-  .magicWand {
-    display: block;
-    margin: 0 auto 20px auto;
-    animation: wandMove 6s infinite ease-in-out;
-  }
-}
