+++ components/app-sidebar.tsx
@@ -26,7 +26,7 @@ const data = {
   navMain: [
     {
       title: "Dashboard",
-      url: "/dashboard",
+      url: "/",
       icon: Home,
       isActive: true,
