+++ components/deleteKey/DeleteKey.tsx
@@ -1,11 +1,11 @@
 "use client"
 
-import { deleteApiKeyProxy } from "@/actions/proxy"
 import { Button } from "../ui/button"
 import { RotateCcwIcon, Trash } from "lucide-react"
 import { EMPTY_FORM_STATE } from "@/lib/zodErrorHandle"
 import { toast } from "@/hooks/use-toast"
 import { useFormStatus } from "react-dom"
+import { deleteApiKey } from "@/actions/apiKeys"
 
 export const DeleteKey = ({ apikey }: { apikey: string }) => {
   const { pending } = useFormStatus();
@@ -14,7 +14,7 @@ export const DeleteKey = ({ apikey }: { apikey: string }) => {
     action={async (form) => {
       try {
         form.append("apikey", apikey)
-        const current = await deleteApiKeyProxy(EMPTY_FORM_STATE, form)
+        const current = await deleteApiKey(EMPTY_FORM_STATE, form)
         if (current.status !== 'SUCCESS') {
           throw new Error(current.message)
         }
