+++ components/nav-user.tsx
@@ -44,6 +44,7 @@ export function NavUser({
   const { isMobile } = useSidebar()
   const { setTheme } = useTheme()
   const { push } = useRouter()
+  const customerPortal = process.env.NEXT_PUBLIC_PADDLE_CUSTOMER_PORTAL_URL
 
   return (
     <SidebarMenu>
@@ -84,10 +85,10 @@ export function NavUser({
               </div>
             </DropdownMenuLabel>
             <DropdownMenuGroup>
-              <DropdownMenuItem onClick={() => push(process.env.NEXT_PUBLIC_PADDLE_CUSTOMER_PORTAL_URL!)} className="cursor-pointer" >
+              {customerPortal && <DropdownMenuItem onClick={() => push(customerPortal)} className="cursor-pointer" >
                 <CreditCard />
                 Billing
-              </DropdownMenuItem>
+              </DropdownMenuItem>}
               <DropdownMenuItem onClick={() => setTheme(curr => curr === "dark" ? "light" : "dark")} className="cursor-pointer" >
                 <Sun className="h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
                 <Moon className="absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
